+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Register build plugin to prepare V2 metadata (enabled via adapter option `build: true`)
 - Skip conversion for element name only message targets
 - Replace deprecated `substr` usages
+- Check ModelProviderService before access
 
 ## Version 1.11.9 - 2023-11-20